requirements.txt
Django==5.1.5
djangorestframework==3.15.2
drf_yasg==1.21.8
nanoid==2.0.0
django-payments==3.0.1
django-phonenumber-field==8.0.0
django-background-tasks==1.2.8
django-compat==1.0.15
django-storages==1.14.4
pytest
pytest-django
pytest-cov
